daily-hud/requirements.txt
requests>=2.28.0
python-whois>=0.8.0
PyYAML>=6.0
rich>=13.0.0
